Change to a file-per-covered-function format
